// src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const AdminRole = {
  SUPER_ADMIN: 'SUPER_ADMIN',
  ADMIN: 'ADMIN'
} as const

export type AdminRole = (typeof AdminRole)[keyof typeof AdminRole]


export const SubmissionStatus = {
  NEW: 'NEW',
  READ: 'READ',
  ARCHIVED: 'ARCHIVED'
} as const

export type SubmissionStatus = (typeof SubmissionStatus)[keyof typeof SubmissionStatus]


export const GalleryCategory = {
  EDUCATION: 'EDUCATION',
  HEALTH: 'HEALTH',
  FOOD: 'FOOD',
  GUIDANCE: 'GUIDANCE',
  SHELTER: 'SHELTER',
  SKILLS: 'SKILLS',
  EVENTS: 'EVENTS',
  OTHER: 'OTHER'
} as const

export type GalleryCategory = (typeof GalleryCategory)[keyof typeof GalleryCategory]